Starting Template
I Follow a tutorial from Awesome Tuts on Udemy to make a game called
Cowboy Runner. I started on 3/21/2016  and finished it on 3/30/2016. I
plan to use this as a starting point to create a game called
PaulyBounce, an endless runner type of game. Hopefully this starting
point will be useful!