getting somewhere
I added objects to the scene to test the machcanics unitl I can I get
the sprites for the game. added movement to the buttons, they are not
functioning how I want them to, but it is a start. I edited the Gameplay
script so it will not keep the scoring system that I do not want either.