did stuff, but don't know if it works
I edited a script to add a coin spawner and I can control where they
spawn at in the y axis. made adjustments to the Gameplay Scene to
include the coin placeholder elements( don't have the sprites for them
yet). mad the coin text. and then I got confused about how I am going to
handle the destroying of the coins. So, I stopped there for today and
will start again on Monday and see if everything works so saw.