Set backs, but moving forward
4/12/2016 - I tried to implement a coin collecting system, but that did
not work as plan. So I took out the elements out of the game involving
the coin system, I may attempt putting it back it before release, but
not sure yet.  4/13/2016 worked on the scoring system, got it to work
somewhat. I did not get the bools to work on there own. After that tried
to fix the movement, but that didn't work out.. Finally I made a commit
on the desktop application I have for GitHub. I made the commit to
master before I wrote a description of what I did. I deleted the commit
not knowing it take my project to my most recent commit before that.
After that I got sick for a week and a half, did not do any work while I
was sick.  4/26/2016 I  redid the removal all over again and got rid of
some tags and the ceiling object, and got the movement working properly.